feat: move page to the upper lvl
